% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getnrows}
\alias{getnrows}
\title{Get the number of rows of the file}
\usage{
getnrows(file, n = 10000)
}
\arguments{
\item{file}{the name of a file (possible with a path)}

\item{n}{the size of the chunks used by the iterator}
}
\value{
an integer
}
\description{
Use iterators to avoid the memory overhead of
obtaining the number of rows of a file.
}
\examples{
data(CO2)
write.csv(CO2, "CO2.csv", row.names=FALSE)
getnrows("CO2.csv")
}

