\name{YaleToolkit-package}
\alias{YaleToolkit-package}
\alias{YaleToolkit}
\docType{package}
\title{
  Data exploration tools from the Department of Statistics at Yale University
}
\description{
  This collection of data exploration tools was developed
  at Yale University for the graphical exploration of complex
  multivariate data. The main functions provided are \code{barcode()},
  \code{gpairs()}, \code{whatis()}, and \code{sparkmat()}, although \code{barcode()}
  and \code{gpairs()} are now provided by packages of the same names, respectively.
}
\details{
  \tabular{ll}{
    Package: \tab YaleToolkit\cr
    Type: \tab Package\cr
    Version: \tab 4.1\cr
    Date: \tab 2012-02-19\cr
    License: \tab GPL version 2 or newer\cr
  }
  See individual help files for more information about the functions.

The package also includes several data sets.  For more information, please see the help files for
\code{nasa} and \code{YaleEnergy}.
Please get in touch with us if you note any problems.

}
\author{
  John W. Emerson, Walton Green

  Maintainer: John W. Emerson <john.emerson@yale.edu>
}
\references{
Chambers, J.M., Cleveland, W.S., Kleiner, B., and Tukey, P.A. (1983), \emph{Graphical Methods for Data Analysis}, Belmont, CA: Wadsworth.

Friendly, M. (2002) 'Corrgrams: Exploratory displays for correlation matrices' \code{American Statistician} 56(4), 316--324.

Tufte, Edward R. (2006) \emph{Beautiful Evidence} The Graphics Press, Cheshire, Connecticut.
See \url{http://www.edwardtufte.com} for this and other references.

}
\keyword{package}
\keyword{hplot}
\keyword{multivariate}
\keyword{ts}

\seealso{
  \code{\link{barcode}}, \code{\link{gpairs}}, \code{\link{sparkmat}},
  \code{\link{sparklines}}, \code{\link{sparkline}},
  \code{\link{whatis}}, \code{\link{corrgram}}.
}
\examples{
  # See individual function documentation for examples.
}
